+++ cloudwatch.tf
@@ -5,13 +5,14 @@
  */
 
 resource "aws_cloudwatch_log_group" "waggledance_ecs" {
-  name = local.instance_alias
-  tags = var.tags
+  count = var.wd_instance_type == "ecs" ? 1 : 0
+  name  = local.instance_alias
+  tags  = var.tags
 }
 
 resource "aws_cloudwatch_dashboard" "apiary_federation" {
+  count          = var.wd_instance_type == "ecs" ? 1 : 0
   dashboard_name = "${local.instance_alias}-${var.aws_region}"
-
   dashboard_body = <<EOF
   {
     "widgets": [
@@ -51,16 +52,16 @@ EOF
 locals {
   alerts = [
     {
-      alarm_name = "${local.instance_alias}-cpu"
-      namespace = "AWS/ECS"
+      alarm_name  = "${local.instance_alias}-cpu"
+      namespace   = "AWS/ECS"
       metric_name = "CPUUtilization"
-      threshold = "80"
+      threshold   = "80"
     },
     {
-      alarm_name = "${local.instance_alias}-memory"
-      namespace = "AWS/ECS"
+      alarm_name  = "${local.instance_alias}-memory"
+      namespace   = "AWS/ECS"
       metric_name = "MemoryUtilization"
-      threshold = "70"
+      threshold   = "70"
     },
   ]
 
@@ -77,18 +78,18 @@ locals {
 }
 
 resource "aws_cloudwatch_metric_alarm" "waggledance_alert" {
-  count = length(local.alerts)
-  alarm_name = local.alerts[count.index].alarm_name
+  count               = var.wd_instance_type == "ecs" ? length(local.alerts) : 0
+  alarm_name          = local.alerts[count.index].alarm_name
   comparison_operator = lookup(local.alerts[count.index], "comparison_operator", "GreaterThanOrEqualToThreshold")
-  metric_name = local.alerts[count.index].metric_name
-  namespace = local.alerts[count.index].namespace
-  period = lookup(local.alerts[count.index], "period", "120")
-  evaluation_periods = lookup(local.alerts[count.index], "evaluation_periods", "2")
-  statistic = "Average"
-  threshold = local.alerts[count.index].threshold
+  metric_name         = local.alerts[count.index].metric_name
+  namespace           = local.alerts[count.index].namespace
+  period              = lookup(local.alerts[count.index], "period", "120")
+  evaluation_periods  = lookup(local.alerts[count.index], "evaluation_periods", "2")
+  statistic           = "Average"
+  threshold           = local.alerts[count.index].threshold
 
   #alarm_description         = ""
   insufficient_data_actions = []
-  dimensions = local.dimensions[count.index]
-  alarm_actions = [aws_sns_topic.apiary_federation_ops_sns.arn]
+  dimensions                = local.dimensions[count.index]
+  alarm_actions             = [aws_sns_topic.apiary_federation_ops_sns[0].arn]
 }