+++ variables.tf
@@ -11,7 +11,7 @@ variable "instance_name" {
 }
 
 variable "wd_instance_type" {
-  description = "Waggle Dance instance type, possible values: ecs,ec2."
+  description = "Waggle Dance instance type, possible values: ecs,k8s."
   type        = "string"
   default     = "ecs"
 }
@@ -20,12 +20,6 @@ variable "waggledance_version" {
   description = "Waggle Dance version to install on EC2 nodes."
   type        = "string"
   default     = "3.3.2"
-}
-
-variable "ami_id" {
-  description = "Amazon Linux AMI, when using ec2 instance type for Waggle Dance."
-  type        = "string"
-  default     = ""
 }
 
 variable "root_vol_type" {
@@ -40,16 +34,10 @@ variable "root_vol_size" {
   default     = "10"
 }
 
-variable "ec2_instance_type" {
-  description = "Waggle Dance EC2 instance type."
+variable "k8s_namespace" {
+  description = "K8s namespace to create waggle-dance deployment."
   type        = "string"
-  default     = "m5.large"
-}
-
-variable "key_name" {
-  description = "EC2 key pair name."
-  type        = "string"
-  default     = "automation"
+  default     = "metastore"
 }
 
 variable "aws_region" {
@@ -91,7 +79,7 @@ The amount of memory (in MiB) used to allocate for the Waggle Dance container.
 Valid values: https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task-cpu-memory-error.html
 EOF
 
-  type = "string"
+  type    = "string"
   default = "4096"
 }
 
@@ -119,6 +107,12 @@ variable "docker_image" {
 variable "docker_version" {
   description = "Waggle Dance Docker image version."
   type        = "string"
+}
+
+variable "k8s_docker_registry_secret" {
+  description = "Docker Registry authentication K8s secret name."
+  type        = "string"
+  default     = ""
 }
 
 variable "graphite_host" {
@@ -185,30 +179,30 @@ The secret's key should be `private_key` and the value should be stored as a bas
 Max character limit for a secret's value is 4096.
 EOF
 
-  type = "string"
+  type    = "string"
   default = ""
 }
 
 variable "enable_remote_metastore_dns" {
   description = "Option to enable creating DNS records for remote metastores."
-  type = "string"
-  default = ""
+  type        = "string"
+  default     = ""
 }
 
 variable "domain_extension" {
   description = "Domain name to use for Route 53 entry and service discovery."
-  type = "string"
-  default = "lcl"
+  type        = "string"
+  default     = "lcl"
 }
 
 variable "secondary_vpcs" {
   description = "List of VPCs to associate with Service Discovery namespace."
-  type = "list"
-  default = []
+  type        = "list"
+  default     = []
 }
 
 variable "docker_registry_auth_secret_name" {
   description = "Docker Registry authentication SecretManager secret name."
-  type = "string"
-  default = ""
+  type        = "string"
+  default     = ""
 }