+++ CHANGELOG.md
@@ -3,13 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [3.0.0] - TBD
+## [3.0.0] - 2019-11-12
 
 ### Added
+- Support for running Waggle Dance on Kubernetes.
 - Upgrade to Terraform version 0.12
 
 ### Changed
 - Tag remote metastore VPC endpoints.
+
+### Removed
+- Support for running Waggle Dance on EC2 nodes.
 
 ## [2.0.1] - 2019-07-17